+++ frontend/src/components/PermissionConfig/index.js
