+++ frontend/src/views/index.js
@@ -1,9 +1,17 @@
-import Admin from "./Admin";
 import Footer from "./Footer";
 import FrameworkDocView from "./FrameworkDocView";
 import Hazard from "./Hazard";
 import Home from "./Home";
+import PermissionConfig from "./PermissionConfig";
 import Profile from "./Profile";
 import Project from "./Project";
 
-export { Admin, Footer, FrameworkDocView, Hazard, Home, Profile, Project };
+export {
+  Footer,
+  FrameworkDocView,
+  Hazard,
+  Home,
+  PermissionConfig,
+  Profile,
+  Project,
+};
